Make proposition dragging easier
This allows propositions to be dragged by a clicking anywhere on the
main panel, including text and whitespace.

Selecting text is now more difficult, but can be done by clicking
outside the proposition and dragging your selection inside.

If this is a significant issue, a "Copy text" action might be
worthwhile.